+++ CHANGELOG.md
@@ -1,10 +1,11 @@
-# 0.0.9 - xx.01.2025
+# 0.0.9 - 25.02.2025
 
 - Lagt til select-komponent
 - Lagt til typografi
 - Lagt til toast-komponent
 - Endret verdien på propen "prioritet" på alert-komponenten fra 'Info' | 'Danger' | 'Warning' | 'Success' til 'info' | 'danger' | 'warning' | 'success'
 - Fjernet default header verdi på alert-komponenten.
+- Lagt til popover-komponent
 
 # 0.0.8 - 13.01.2025